Migrating to CSV
- Migrated the random scheduler to its own module that uses 'scheduler.py' as the driver
- Included the csv file that contains the routes
