+++ src/config/general.yaml
@@ -8,7 +8,7 @@ plot: 1
 time_limit: 100
 
 # Output previous decision variable/param results
-load_from_file: 1
+load_from_file: 0
 
 # Load previous input parameters
 run_prev: 0
