+++ src/config/schedule.yaml
@@ -2,15 +2,15 @@ time:
   time_horizon: 24     # [hr]
   K: 1000              # [steps]
 buses:
-  num_bus: 20
-  num_visit: 250
+  num_bus: 15
+  num_visit: 200
   bus_length: 1
   bat_capacity: 387.78 # [KWh]
   max_charge: 1        # [0-1]
   min_charge: 0.25     # [0-1]
-  dis_rate: 5          # [Kw]
+  dis_rate: 4          # [Kw]
   max_rest: 0.5       # [hr]
-  min_rest: 0.25       # [hr]
+  min_rest: 0.15       # [hr]
 initial_charge:
   max: 1               # [0-1]
   min: 1               # [0-1]
